Implementation of the BlurEffect
