+++ src/Runtime/Runtime/Runtime.OpenSilver.csproj
@@ -308,7 +308,7 @@
 		<Compile Include="System.Windows.Browser\WORKINPROGRESS\HtmlDocument.cs" />
 		<Compile Include="System.Windows.Browser\WORKINPROGRESS\HtmlPopupWindowOptions.cs" />
 		<Compile Include="System.Windows.Media.Animation\PowerEase.cs" />
-		<Compile Include="System.Windows.Media.Effects\WORKINPROGRESS\BlurEffect.cs" />
+		<Compile Include="System.Windows.Media.Effects\BlurEffect.cs" />
 		<Compile Include="System.Windows.Messaging\WORKINPROGRESS\ListenFailedException.cs" />
 		<Compile Include="System.Windows.Messaging\WORKINPROGRESS\LocalMessageReceiver.cs" />
 		<Compile Include="System.Windows.Messaging\WORKINPROGRESS\LocalMessageSender.cs" />
